Refactor: Move users gathering code to UsersGathering
