+++ lib/models/tencent_agent/users_gathering.rb
@@ -1,7 +1,6 @@
 class TencentAgent
   module UsersGathering
     extend ActiveSupport::Concern
-    include TweetsGathering
 
     KEYWORDS_QUEUE = 'spider:tencent:users_gathering:keywords_queue'
 
@@ -33,6 +32,73 @@ class TencentAgent
 
     rescue Error => e
       $logger.err log("Aborted users gathering: #{e.message}")
+    end
+
+    private
+
+    def try_publish_user(user_name)
+      result = cached_get('api/user/other_info', name: user_name)
+
+      if result['ret'].zero? && result['data']
+        user = UserFilter.filter(result['data'])
+
+        if user
+          group_ids = get_group_ids(user)
+
+          unless group_ids.empty?
+            publish_user(user)
+            group_ids.each do |group_id|
+              publish_user_to_group(user, group_id)
+            end
+            return true
+          end
+        else
+          $logger.notice log(%{Skip invalid user "#{user_name}"})
+        end
+
+      else
+        $logger.err log(%{Failed to gather profile of user "#{user_name}"})
+      end
+      false
+    end
+
+    def publish_user(user)
+      $logger.notice log(%{Publishing user "#{user['name']}"})
+      $redis.rpush 'streaming/messages', {
+        type: 'add_user',
+        body: {
+          id: user['name'],
+          type: 'tencent',
+          birth_year: user['birth_year'],
+          gender: user['gender'],
+          city: user['city']
+        }
+      }.to_json
+    end
+
+    def get_group_ids(user)
+      response = Faraday.get(
+        GET_GROUP_IDS_URL,
+        birth_year: user['birth_year'],
+        city: user['city'],
+        gender: user['gender']
+      )
+      MultiJson.load(response.body)['ids']
+    rescue
+      $logger.err log(%{Failed to get group ids for user "#{user['name']}"})
+      []
+    end
+
+    def publish_user_to_group(user, group_id)
+      $logger.notice log(%{Publishing user "#{user['name']}" to group "#{group_id}"})
+      $redis.rpush 'streaming/messages', {
+        type: 'add_user_to_group',
+        body: {
+          group_id: group_id,
+          user_id: user['name'],
+          user_type: 'tencent'
+        }
+      }.to_json
     end
   end
 end
